+++ dev/cmb/simulation-workflows/ats/internal/tests/test_coordinator/test_coordinator.py
@@ -1,45 +1,15 @@
 import os
-import sys
 import unittest
 
-import smtk
-import smtk.attribute
-import smtk.io
-import smtk.operation
-import smtk.resource
-import smtk.session.vtk
-
 from writer import ats_writer, coordinator_writer
+
+from tests.base import BaseTestCase
 
 ATT_RESOURCE_FILENAME = 'att.coordinator.smtk'
 # BASLINE_XML_FILENAME = 'baseline_vis.xml'
 
-OPERATION_SUCCEEDED = int(smtk.operation.Operation.SUCCEEDED)  # 3
 
-class CoordinatorTest(unittest.TestCase):
-
-    def setUp(self):
-        self.att_resource = None
-        self.model_resource = None
-
-        # Initialize smtk managers
-        self.res_manager = smtk.resource.Manager.create()
-        self.op_manager = smtk.operation.Manager.create()
-
-        smtk.attribute.Registrar.registerTo(self.res_manager)
-        smtk.attribute.Registrar.registerTo(self.op_manager)
-
-        smtk.session.vtk.Registrar.registerTo(self.res_manager)
-        smtk.session.vtk.Registrar.registerTo(self.op_manager)
-
-        smtk.operation.Registrar.registerTo(self.op_manager)
-        self.op_manager.registerResourceManager(self.res_manager)
-
-    def tearDown(self):
-        self.res_manager = None
-        self.op_manager = None
-        self.att_resource = None
-        self.model_resource = None
+class CoordinatorTest(BaseTestCase):
 
     def test_coordinator(self):
         """"""
@@ -63,16 +33,6 @@ class CoordinatorTest(unittest.TestCase):
         # with open(baseline_path) as fp:
         #     baseline_string = fp.read()
         # self.assertEqual(xml_string, baseline_string)
-
-    def _read_resource(self, path):
-        read_op = self.op_manager.createOperation('smtk::operation::ReadResource')
-        read_op.parameters().find('filename').setValue(path)
-        read_result = read_op.operate()
-        read_outcome = read_result.findInt('outcome').value(0)
-        self.assertEqual(read_outcome, OPERATION_SUCCEEDED)
-        resource = read_result.find('resource').value()
-        self.assertIsNotNone(resource)
-        return resource
 
 
 if __name__ == '__main__':
